+++ src/applications/vaos/sass/vaos.scss
@@ -10,7 +10,7 @@
 
   .vaos-appts__tab {
     width: 50%;
-    font-size:14px;
+    font-size: 14px;
     border-color: $color-gray-light;
     border: solid 1px $color-gray-light;
     border-radius: 0 !important;
@@ -31,10 +31,6 @@
       width: 200px;
       font-size: 16px;
     }
-  }
-
-  @media only screen and (max-width: 480px) {
-    
   }
 }
